Replace event subscribers with search api processors
